Move mcpe/ to src/minecraft/
