+++ src/main.cpp
@@ -15,17 +15,17 @@
 #include "android_symbols.h"
 #include "egl_symbols.h"
 #include "fmod_symbols.h"
-#include "../mcpe/gl.h"
-#include "../mcpe/AppPlatform.h"
-#include "../mcpe/MinecraftGame.h"
+#include "minecraft/gl.h"
+#include "minecraft/AppPlatform.h"
+#include "minecraft/MinecraftGame.h"
 #include "linux_appplatform.h"
 #include "linux_store.h"
-#include "../mcpe/Mouse.h"
-#include "../mcpe/Keyboard.h"
-#include "../mcpe/Options.h"
+#include "minecraft/Mouse.h"
+#include "minecraft/Keyboard.h"
+#include "minecraft/Options.h"
 #include "common.h"
 #include "hook.h"
-#include "../mcpe/Xbox.h"
+#include "minecraft/Xbox.h"
 #include "browser.h"
 #include "xboxlive.h"
