+++ src/minecraft/ImagePickingCallback.h
