+++ src/minecraft/string.cpp
