+++ src/minecraft/string.h
