+++ src/xboxlive.h
@@ -2,7 +2,7 @@
 
 #include "msa.h"
 #include "cll.h"
-#include "../mcpe/Xbox.h"
+#include "minecraft/Xbox.h"
 #include <map>
 
 class SimpleMSAStorageManager;
